move fusions
